:construction: add opcode gen

+++ src/plinf.yy.c
@@ -1192,7 +1192,7 @@ YY_RULE_SETUP
 {
   CONST *temp = (CONST *)calloc(1, sizeof(CONST));
   temp->type = ct_int;
-  temp->num = atoi(yytext);
+  temp->num = atol(yytext);
   yylval.const_value = temp;
   return NUMBER;
 }
